Fit validation for floating point values

+++ heron_of_alexandria/logic.py
@@ -1,7 +1,7 @@
 def validate(val):
 
     try:
-        _temp = int(val)
+        _temp = float(val)
     except ValueError:
         return False
